strip cdot and degree from bases that become forms

+++ misc/l2phase3/l2p3-xff.sh
@@ -1,9 +1,22 @@
 #!/bin/sh
+echo xff=yes
 webdir=$1
 l=$2
-echo xff=yes
+if [ "$webdir" == "" ]; then
+    echo "$0: usage: WEBDIR LANG. Stop"
+    exit 1
+fi
+if [ "$l" == "" ]; then
+    echo "$0: usage: WEBDIR LANG. Stop"
+    exit 1
+fi
+if [ ! -r 01bld/$l/articles.xml ]; then
+    echo "$0: 01bld/$l/articles.xml does not exist. Stop"
+    exit 1
+fi
 xffdir=$webdir/cbd/$l/xff
 rm -fr $xffdir
+mkdir -p $xffdir
 art2xff.plx -d $xffdir -l sux
 chmod +rx $webdir/cbd/$l/xff
 chmod -R o+r $webdir/cbd/$l/xff
